Refactored Plotting OUT of PortWatcher object
